4.4 Backport Add config setting to disable mandatory tokens check in mailings
As per a previous discussion (with Lobo IIRC), add a configuration
setting to disable the check for presence of mandatory tokens when
sending mailings.

This setting is implemented as a negative option, i.e. ticking the
option disables the check. We could also reverse the logic, so that
ticking the option enables the check instead; and set it to on by
default... Not sure which is preferable -- I guess that's mostly a
matter of taste.

Conflicts:

	CRM/Admin/Form/Preferences/Mailing.php
	settings/Mailing.setting.php