CRM-11551: fix support for time in Report date ranges
Time field information was being stripped from the date in some places.

There was a 12/24-hour bug in parsing the time, which would display an
endpoint of midnight as noon in the criteria summary.

When the time field is present, we were incorrectly interpreting a blank value
to mean midnight on both ends, when the correct logic is to assume 23:59:59
at the later endpoint.

At WMF, we have an additional patch hunk which will always display time in the
report criteria summary:
-            $from = CRM_Utils_Date::customFormat($from, NULL, array('d'));
-            $to = CRM_Utils_Date::customFormat($to, NULL, array('d'));
+            $from = CRM_Utils_Date::customFormat($from, '%B %E%f, %Y %H:%Mh');
+            $to = CRM_Utils_Date::customFormat($to, '%B %E%f, %Y %H:%Mh');